requirements.txt
numpy~=2.2.4
torch~=2.6.0
scipy~=1.15.2
python_Levenshtein~=0.27.1
transformers~=4.50.3
evaluate~=0.4.3
sacrebleu~=2.5.1
